switch to having 2 different functions for sort() and sort-by()
