+++ src/main/java/org/javarosa/xpath/expr/FunctionUtils.java
@@ -81,6 +81,7 @@ public class FunctionUtils {
         funcList.put(XPathIdCompressFunc.NAME, XPathIdCompressFunc.class);
         funcList.put(XPathJoinChunkFunc.NAME, XPathJoinChunkFunc.class);
         funcList.put(XPathSortFunc.NAME, XPathSortFunc.class);
+        funcList.put(XPathSortFunc.NAME, XPathSortFunc.class);
     }
 
     private static final CacheTable<String, Double> mDoubleParseCache = new CacheTable<>();
