+++ src/main/java/org/javarosa/xpath/parser/ast/ASTNodeFunctionCall.java
@@ -165,6 +165,8 @@ public class ASTNodeFunctionCall extends ASTNode {
                 return new XPathIdCompressFunc(args);
             case "sort":
                 return new XPathSortFunc(args);
+            case "sort-by":
+                return new XPathSortByFunc(args);
             default:
                 return new XPathCustomRuntimeFunc(name, args);
         }
